+++ schedulerx3-20240624/alibabacloud_schedulerx320240624/models/_create_job_request.py
@@ -43,11 +43,11 @@ class CreateJobRequest(DaraModel):
         # 
         # This parameter is required.
         self.app_name = app_name
-        # The retry interval on failure. Unit: seconds. Default value: 30.
+        # The retry interval upon failure. Unit: seconds. Default value: 30.
         self.attempt_interval = attempt_interval
-        # The custom calendar. This parameter is optional for the cron time type.
+        # The custom calendar. This parameter is available only for the cron time type.
         self.calendar = calendar
-        # The child node IDs, separated by commas.
+        # The child node IDs. Separate multiple IDs with commas.
         self.child_job_id = child_job_id
         # The cluster ID.
         # 
@@ -59,10 +59,10 @@ class CreateJobRequest(DaraModel):
         self.dependent_strategy = dependent_strategy
         # The node description.
         self.description = description
-        # The client blocking strategy. Valid values:
+        # The client-side blocking strategy. Valid values:
         # - 1: serial execution on a single machine
-        # - 2: ignore subsequent schedules
-        # - 3: override previous schedules.
+        # - 2: ignore subsequent scheduling
+        # - 3: override previous scheduling
         self.executor_block_strategy = executor_block_strategy
         # The JobHandler name.
         self.job_handler = job_handler
@@ -70,7 +70,7 @@ class CreateJobRequest(DaraModel):
         # 
         # This parameter is required.
         self.job_type = job_type
-        # The maximum number of retries on failure. Set this parameter based on your business requirements.
+        # The maximum number of retry attempts upon failure. Set this parameter based on your business requirements.
         self.max_attempt = max_attempt
         # The maximum number of concurrent instances.
         self.max_concurrency = max_concurrency
@@ -88,7 +88,7 @@ class CreateJobRequest(DaraModel):
         # - 1: low
         # - 5: medium
         # - 10: high
-        # - 15: very high.
+        # - 15: very high
         self.priority = priority
         # The routing strategy. Valid values:
         # - 1: round robin
@@ -98,7 +98,7 @@ class CreateJobRequest(DaraModel):
         # - 5: least frequently used
         # - 6: least recently used
         # - 7: consistent hashing
-        # - 8: shard broadcast.
+        # - 8: shard broadcast
         self.route_strategy = route_strategy
         # The script content for non-BEAN node types. Use this field to pass the script content.
         self.script = script
@@ -108,21 +108,21 @@ class CreateJobRequest(DaraModel):
         self.start_time_type = start_time_type
         # The node status. Default value: enabled. Valid values:
         # - 0: disabled
-        # - 1: enabled.
+        # - 1: enabled
         self.status = status
         # The time expression. Set this parameter based on the selected time type.
-        # - **none**: No value is required.
-        # - **cron**: Specify a standard cron expression. Online validation is supported.
-        # - **api**: No value is required.
-        # - **fixed_rate**: Specify a fixed frequency value in seconds. For example, 200 indicates that the node is triggered every 200 seconds.
-        # - **one_time**: Specify a scheduling time in the yyyy-MM-dd HH:mm:ss format or a timestamp in milliseconds. For example, "2022-10-10 10:10:00".
+        # - **none**: no value required.
+        # - **cron**: a standard cron expression. Online verification is supported.
+        # - **api**: no value required.
+        # - **fixed_rate**: a fixed frequency value in seconds. For example, 200 indicates that the node is triggered every 200 seconds.
+        # - **one_time**: a scheduling time in the yyyy-MM-dd HH:mm:ss format or a timestamp in milliseconds. For example, "2022-10-10 10:10:00".
         self.time_expression = time_expression
         # The time type. Valid values:
         # - -1: none<br/>
         # - 1: cron<br/>
         # - 3: fix_rate<br/>
         # - 5: one_time<br/>
-        # - 100: api.
+        # - 100: api
         # 
         # This parameter is required.
         self.time_type = time_type
@@ -132,7 +132,7 @@ class CreateJobRequest(DaraModel):
         self.weight = weight
         # The extended attributes. This parameter is required for K8s node types.
         # Job node: {"resource":"job"}
-        # Shell node: {"image":"busybox","resource":"shell"}.
+        # Shell node: {"image":"busybox","resource":"shell"}
         self.xattrs = xattrs
 
     def validate(self):
@@ -337,11 +337,11 @@ class CreateJobRequestNoticeContacts(DaraModel):
         contact_type: int = None,
         name: str = None,
     ):
-        # The object type of the notification recipient. Valid values:
+        # The Notification Recipient type. Valid values:
         # 
         # - 1: alert contact
         # 
-        # - 2: alert contact group.
+        # - 2: alert contact group
         self.contact_type = contact_type
         # The name of the alert contact or alert contact group.
         self.name = name
@@ -390,18 +390,18 @@ class CreateJobRequestNoticeConfig(DaraModel):
         self.end_early_enable = end_early_enable
         # Specifies whether to enable failure alerting. Valid values:
         # 
-        # - **true**: Enabled.
-        # - **false**: Disabled.
+        # - **true**: Failure alerting is enabled.
+        # - **false**: Failure alerting is disabled.
         self.fail_enable = fail_enable
         # The number of consecutive failures.
         # > An alert is sent only when the number of consecutive failures exceeds the configured value.
         self.fail_limit_times = fail_limit_times
         # Specifies whether to enable alerting when no workers are available. Valid values:
-        # - **true**: Enabled.
-        # - **false**: Disabled.
+        # - **true**: Alerting for no available workers is enabled.
+        # - **false**: Alerting for no available workers is disabled.
         self.miss_worker_enable = miss_worker_enable
         # The notification channel. Valid values:
-        # - sms: SMS
+        # - sms: text message
         # - phone: phone call
         # - mail: email
         # - webhook: webhook
@@ -422,8 +422,8 @@ class CreateJobRequestNoticeConfig(DaraModel):
         self.timeout_enable = timeout_enable
         # Specifies whether to enable timeout termination. Valid values:
         # 
-        # - **true**: Enabled.
-        # - **false**: Disabled.
+        # - **true**: Timeout termination is enabled.
+        # - **false**: Timeout termination is disabled.
         self.timeout_kill_enable = timeout_kill_enable
 
     def validate(self):