+++ schedulerx3-20240624/alibabacloud_schedulerx320240624/models/_delete_contact_response_body.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from darabonba.model import DaraModel
 
-class UpdateJobResponseBody(DaraModel):
+class DeleteContactResponseBody(DaraModel):
     def __init__(
         self,
         code: int = None,
@@ -12,15 +12,9 @@ class UpdateJobResponseBody(DaraModel):
         request_id: str = None,
         success: bool = None,
     ):
-        # The response code.
         self.code = code
-        # The error message.
         self.message = message
-        # The request ID. Alibaba Cloud generates a unique identifier for each request. You can use the request ID to troubleshoot issues.
         self.request_id = request_id
-        # Indicates whether the call was successful. Valid values:
-        # - true: The call was successful.
-        # - false: The call failed.
         self.success = success
 
     def validate(self):