+++ schedulerx3-20240624/alibabacloud_schedulerx320240624/models/_update_job_shrink_request.py
@@ -38,118 +38,88 @@ class UpdateJobShrinkRequest(DaraModel):
         # 
         # This parameter is required.
         self.app_name = app_name
-        # The interval in seconds between retry attempts.
+        # The retry interval upon node failure.
         self.attempt_interval = attempt_interval
         # The custom calendar.
         self.calendar = calendar
-        # The client blocking strategy.
-        # 
-        # - 1: Serial execution
-        # 
-        # - 2: Ignore later schedules
-        # 
-        # - 3: Overwrite earlier schedules
+        # The child node IDs, separated by commas.
         self.child_job_id = child_job_id
         # The cluster ID.
         # 
         # This parameter is required.
         self.cluster_id = cluster_id
+        # The dependency strategy.
         self.dependent_strategy = dependent_strategy
-        # The job description.
+        # The node description.
         self.description = description
-        # Notification contact configuration
+        # The client blocking strategy. Valid values:
+        # - 1: serial execution on a single machine
+        # - 2: ignore subsequent scheduling
+        # - 3: override previous scheduling
         self.executor_block_strategy = executor_block_strategy
-        # The job handler name.
+        # The JobHandler name.
         self.job_handler = job_handler
-        # The job ID.
+        # The node ID.
         # 
         # This parameter is required.
         self.job_id = job_id
-        # The maximum number of retries for a failed job.
+        # The maximum number of retry attempts upon node failure.
         self.max_attempt = max_attempt
-        # The maximum number of concurrent job instances.
-        # 
-        # > This parameter defines the maximum number of instances for a single job that can run concurrently. A value of `1` prevents duplicate execution. If this limit is exceeded, the scheduler skips the current job.
+        # The maximum number of concurrent instances of the node.
+        # >The maximum number of instances that can run at the same time for the same node. A value of 1 indicates that repeated execution is not allowed. If the concurrency limit is exceeded, the current scheduling is skipped.
         self.max_concurrency = max_concurrency
-        # The job name.
+        # The node name.
         self.name = name
-        # Time zone
-        # 
-        # > The default is the time zone of the SchedulerX server.
+        # The notification configuration.
         self.notice_config_shrink = notice_config_shrink
-        # Notification configuration
+        # The notification contact configuration.
         self.notice_contacts_shrink = notice_contacts_shrink
-        # The job parameters.
+        # The node parameters.
         self.parameters = parameters
-        # The job execution priority. Valid values:
+        # The execution priority of the node. Valid values:
         # 
-        # - `1`: Low
-        # 
-        # - `5`: Medium
-        # 
-        # - `10`: High
-        # 
-        # - `15`: Very High
+        # - 1: low
+        # - 5: medium
+        # - 10: high
+        # - 15: very high
         self.priority = priority
-        # The routing strategy. Valid values:
+        # The routing policy. Valid values:
         # 
-        # - `1`: round-robin
-        # 
-        # - `2`: random
-        # 
-        # - `3`: first
-        # 
-        # - `4`: last
-        # 
-        # - `5`: least frequently used
-        # 
-        # - `6`: least recently used
-        # 
-        # - `7`: consistent hashing
-        # 
-        # - `8`: sharded broadcast
+        # - 1: round robin
+        # - 2: random
+        # - 3: first
+        # - 4: last
+        # - 5: least frequently used
+        # - 6: least recently used
+        # - 7: consistent hashing
+        # - 8: shard broadcast
         self.route_strategy = route_strategy
-        # The script content for non-BEAN jobs.
+        # The script for non-BEAN nodes. Use this field to configure the script.
         self.script = script
-        # The type of the start time.
+        # The scheduling start time.
         self.start_time = start_time
-        # The task execution priority. The following values are supported:
-        # 
-        # - 1: Low
-        # 
-        # - 5: Medium
-        # 
-        # - 10: High
-        # 
-        # - 15: Very High
+        # The start time type.
         self.start_time_type = start_time_type
-        # The time expression. The expression format depends on the `TimeType`.
+        # The time expression. Set the time expression based on the selected time type.
         # 
-        # - `none`: Leave this parameter empty.
-        # 
-        # - `cron`: Specify a standard cron expression. Online validation is supported.
-        # 
-        # - `api`: Leave this parameter empty.
-        # 
-        # - `fixed_rate`: An integer that represents a fixed interval in seconds. For example, `30` triggers the job every 30 seconds.
-        # 
-        # - `one_time`: A single execution time, specified in the `yyyy-MM-dd HH:mm:ss` format or as a timestamp in milliseconds. For example, "2022-10-10 10:10:00".
+        # - none: No value is required.
+        # - cron: Specify a standard cron expression. Online verification is supported.
+        # - api: No value is required.
+        # - fixed_rate: Specify a fixed frequency value in seconds. For example, 30 indicates that the node is triggered every 30 seconds.
+        # - one_time: Specify a scheduling time in the yyyy-MM-dd HH:mm:ss format or a timestamp in milliseconds. For example, "2022-10-10 10:10:00".
         self.time_expression = time_expression
         # The time type. Valid values:
         # 
-        # - `-1`: none
-        # 
-        # - `1`: cron
-        # 
-        # - `3`: fixed_rate
-        # 
-        # - `5`: one_time
-        # 
-        # - `100`: api
+        # - -1: none
+        # - 1: cron
+        # - 3: fix_rate
+        # - 5: one_time
+        # - 100: api
         self.time_type = time_type
-        # The start time of the schedule.
+        # The time zone.
+        # > By default, the time zone of the SchedulerX server is used.
         self.timezone = timezone
-        # The ID of the child job. Separate multiple IDs with a comma.
+        # The node weight.
         self.weight = weight
         self.xattrs = xattrs